+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# v2.0.1
+## 12/07/2018
+
+1. [](#new)
+    * Added support for admin reporting available in Grav 1.6
+1. [](#bugfix)
+    * Fixed issue with twig auto-escaping
+    * Fixed problems plugin potentially breaking CLI command if plugins get initialized
+
 # v2.0.0
 ## 09/30/2018